sketch-to-sky-backend/requirements.txt
fastapi
uvicorn
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
mlcroissant
trimesh
python-dotenv
google-cloud-storage

# HTTP Client for remote API calls
httpx>=0.27.0
